Some refactoring changes

+++ app/src/main/java/com/oia/ilkan/blg456eproject/Data.java
@@ -13,5 +13,5 @@ public class Data {
 
     public double angle;
     public double power;
-    public int joystick;
+    public byte joystick;
 }
